backend/requirements.txt
Flask
python-dotenv
requests
google-generativeai
psycopg2-binary
Pillow 